refactor(dashboard): components files restructured into 3 main directories
