+++ dashboard/__tests__/Home.test.tsx
@@ -1,7 +1,7 @@
 
 jest.mock('jose');
 import { render, screen } from '@testing-library/react';
-import Home from '../src/components/Home';
+import Home from '../src/components/modules/home/Home';
 
 
 test('renders project description text', () => {
