+++ dashboard/src/app/alerts/alerts-table/page.tsx
@@ -1,6 +1,6 @@
 import { Metadata } from 'next';
-import Alerts from '../../../components/alerts/Alerts';
-import AlertsTabs from '../../../components/alerts/AlertsTabs';
+import Alerts from '../../../components/modules/alerts/Alerts';
+import AlertsTabs from '../../../components/modules/alerts/AlertsTabs';
 export const metadata: Metadata = {
   title: 'OceanBoards - Alerts',
   description: 'AMRIT Alerts dashboard',
