+++ dashboard/src/app/alerts/page.tsx
@@ -1,5 +1,5 @@
 import { Metadata } from 'next';
-import Alerts from '../../components/alerts/Alerts';
+import Alerts from '../../components/modules/alerts/Alerts';
 
 export const metadata: Metadata = {
   title: 'OceanBoards - Alerts',
