+++ dashboard/src/app/login/page.tsx
@@ -1,4 +1,4 @@
-import LoginForm from '@/components/login/LoginForm'
+import LoginForm from '@/components/layout/login/LoginForm'
 import { Metadata } from 'next';
 import React from 'react'
