+++ dashboard/src/components/layout/SectionTabs.tsx
