+++ dashboard/src/components/layout/login/ForgotPassword.tsx
