+++ dashboard/src/components/layout/sidebar/HoverPreview.tsx
